requirements.txt
paho-mqtt
sudo apt-get install libmysqlclient-dev
flask_mysqldb